* VM integration draft

* fix account_as_bytes

* impl publish module

* add tests

* rewrite storage

+++ pallets/sp-mvm/src/mock.rs
@@ -31,7 +31,7 @@ impl system::Trait for Test {
     type BlockNumber = u64;
     type Hash = H256;
     type Hashing = BlakeTwo256;
-    type AccountId = u64;
+    type AccountId = u8;
     type Lookup = IdentityLookup<Self::AccountId>;
     type Header = Header;
     type Event = ();
